Create remote_buttons.js
A particular computerized piece of home electronics operates through your television set. Input is via your wireless remote control and output is displayed on the TV. To input a string of text into the system, a grid of characters is displayed on the screen with a cursor. To input each character of the string, you move the cursor around using the arrow keys on your remote until the cursor is on the desired character, then you press "SELECT". After all the characters are input, you press "ENTER".